Can you try to fix the ones i red commented?
Line 77: if u create a tag with a white space ex.
Project A you cant create a tasks under the tag Project A.
Due to the DOM token not accepting white spaces

Line 205: data.tasks[task_id].Title does not exist
same with the other 2

Note: Do Not add this to main yet medyo delikado pa.

Co-authored-by: jrs-a <jrs-a@users.noreply.github.com>